# backend/node/docs/openapi.yaml
openapi: 3.1.0
info:
  title: GymPoint API
  version: 1.0.0
  description: API del sistema GymPoint para gestión de gimnasios, rutinas, entrenamientos y gamificación.
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo local
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Datos inválidos o request malformada
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Datos inválidos
    Unauthorized:
      description: Autenticación requerida o token inválido/expirado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: No autenticado
    Forbidden:
      description: Permisos insuficientes para realizar la operación
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Acceso denegado
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Recurso no encontrado
...
  schemas:
...
    SubscriptionType:
      type: string
      enum:
        - FREE
        - PREMIUM
      description: Tipo de suscripción del usuario
    Gender:
      type: string
      enum:
        - M
        - F
        - O
      description: Género del usuario (M=Masculino, F=Femenino, O=Otro)
...
    Error:
      type: object
      required:
        - error
      additionalProperties: false
      properties:
        error:
          type: string
          description: Mensaje de error descriptivo
          example: Ha ocurrido un error
...
    UpdateTokensRequest:
      type: object
      required:
        - delta
      additionalProperties: false
      properties:
        delta:
          type: integer
          description: Cantidad a incrementar (positivo) o decrementar (negativo)
          example: 50
        reason:
          type: string
          maxLength: 255
          description: Razón del ajuste de tokens
          example: Recompensa por evento especial
    TokensUpdateResponse:
      type: object
      required:
        - id_user_profile
        - new_balance
        - delta
      additionalProperties: false
      properties:
        id_user_profile:
          type: integer
          description: ID del perfil de usuario
        new_balance:
          type: integer
          minimum: 0
          description: Nuevo balance de tokens
          example: 200
        delta:
          type: integer
          description: Cantidad ajustada
          example: 50
...
    UserProfileResponse:
      type: object
      required:
        - id_user_profile
        - email
        - name
        - lastname
        - subscription
        - tokens
        - created_at
      additionalProperties: false
      properties:
        id_user_profile:
          type: integer
          description: ID único del perfil de usuario
          example: 42
        email:
          type: string
          format: email
          description: Email del usuario
          example: usuario@example.com
        name:
          type: string
          minLength: 2
          maxLength: 80
          description: Nombre del usuario
          example: Juan
        lastname:
          type: string
          minLength: 2
          maxLength: 80
          description: Apellido del usuario
          example: Pérez
        gender:
          oneOf:
            - $ref: '#/components/schemas/Gender'
            - type: 'null'
          description: Género del usuario
          example: M
        birth_date:
          oneOf:
            - type: string
              format: date
            - type: 'null'
          description: Fecha de nacimiento (YYYY-MM-DD)
          example: '1990-05-15'
        locality:
          oneOf:
            - type: string
            - type: 'null'
          description: Localidad del usuario
          example: Buenos Aires
        profile_picture_url:
          oneOf:
            - type: string
              format: uri
            - type: 'null'
          description: URL de la foto de perfil
        subscription:
          allOf:
            - $ref: '#/components/schemas/SubscriptionType'
          description: Nivel de suscripción actual
          example: FREE
        tokens:
          type: integer
          minimum: 0
          description: Balance actual de tokens
          example: 150
        preferred_language:
          type: string
          description: Idioma preferido (código ISO 639-1)
          example: es
        timezone:
          oneOf:
            - type: string
            - type: 'null'
          description: Zona horaria del usuario
          example: America/Argentina/Buenos_Aires
        onboarding_completed:
          type: boolean
          description: Indica si completó el onboarding
          example: true
        created_at:
          type: string
          format: date-time
          description: Fecha de creación del perfil
        updated_at:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
          description: Fecha de última actualización
    UpdateUserProfileRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 80
          description: Nombre del usuario
        lastname:
          type: string
          minLength: 2
          maxLength: 80
          description: Apellido del usuario
        gender:
          $ref: '#/components/schemas/Gender'
        birth_date:
          type: string
          format: date
          description: Fecha de nacimiento (YYYY-MM-DD)
          example: '1990-05-15'
        locality:
          type: string
          maxLength: 100
          description: Localidad del usuario
        profile_picture_url:
          type: string
          format: uri
          description: URL de la foto de perfil
        preferred_language:
          type: string
          pattern: ^[a-z]{2}$
          description: Idioma preferido (código ISO 639-1)
          example: es
        timezone:
          type: string
          description: Zona horaria del usuario
          example: America/Argentina/Buenos_Aires
        onboarding_completed:
          type: boolean
          description: Marca el onboarding como completado
...
    RequestAccountDeletionRequest:
      type: object
      additionalProperties: false
      properties:
        reason:
          type: string
          maxLength: 500
          description: Razón de la eliminación (opcional)
          example: Ya no uso la aplicación
    AccountDeletionResponse:
      type: object
      required:
        - status
        - scheduled_deletion_date
      additionalProperties: false
      properties:
        id_account_deletion_request:
          type: integer
          description: ID de la solicitud de eliminación
        id_account:
          type: integer
          description: ID de la cuenta
        status:
          type: string
          enum:
            - PENDING
            - CANCELLED
            - COMPLETED
          description: Estado de la solicitud
          example: PENDING
        reason:
          oneOf:
            - type: string
            - type: 'null'
          description: Razón proporcionada por el usuario
        scheduled_deletion_date:
          type: string
          format: date-time
          description: Fecha programada para eliminación
        cancelled_at:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
          description: Fecha de cancelación (si aplica)
        metadata:
          type: object
          description: Metadatos adicionales de la solicitud
          properties:
            grace_period_days:
              type: integer
              description: Días de período de gracia
              example: 30
    AccountDeletionStatusResponse:
      type: object
      required:
        - has_active_request
      additionalProperties: false
      properties:
        has_active_request:
          type: boolean
          description: Indica si hay una solicitud activa
          example: true
        request:
          oneOf:
            - $ref: '#/components/schemas/AccountDeletionResponse'
            - type: 'null'
          description: Detalles de la solicitud (si existe)
...
    UpdateSubscriptionRequest:
      type: object
      required:
        - subscription
      additionalProperties: false
      properties:
        subscription:
          type: string
          enum:
            - FREE
            - PREMIUM
          description: Nuevo nivel de suscripción
          example: PREMIUM
...
paths:
...
  /api/users/me:
    get:
      operationId: getUserProfile
      tags:
        - Users
      summary: Obtener perfil del usuario actual
      description: |
        Devuelve la información completa del perfil del usuario autenticado.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Perfil del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: updateUserProfile
      tags:
        - Users
      summary: Actualizar perfil del usuario actual
      description: |
        Actualiza los datos del perfil del usuario autenticado. Solo se actualizan los campos proporcionados en el request body.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserProfileRequest'
      responses:
        '200':
          description: Perfil actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: requestAccountDeletion
      tags:
        - Users
      summary: Solicitar eliminación de cuenta
      description: |
        Inicia el proceso de eliminación de cuenta con período de gracia. La cuenta no se eliminará inmediatamente sino después del período configurado.
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestAccountDeletionRequest'
      responses:
        '200':
          description: Solicitud de eliminación registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDeletionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/users/me/deletion-request:
    get:
      operationId: getAccountDeletionStatus
      tags:
        - Users
      summary: Obtener estado de solicitud de eliminación
      description: |
        Consulta el estado de una solicitud de eliminación de cuenta pendiente.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Estado de la solicitud
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDeletionStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      operationId: cancelAccountDeletion
      tags:
        - Users
      summary: Cancelar solicitud de eliminación
      description: |
        Cancela una solicitud de eliminación de cuenta pendiente.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Solicitud cancelada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDeletionResponse'
        '400':
          description: No hay solicitud pendiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/users/{userId}/tokens:
    post:
      operationId: updateUserTokens
      tags:
        - Users
      summary: Actualizar tokens de un usuario (admin)
      description: |
        Incrementa o decrementa el balance de tokens de un usuario. Solo para administradores.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            minimum: 1
          description: ID del perfil de usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTokensRequest'
      responses:
        '200':
          description: Tokens actualizados exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokensUpdateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/users/{userId}/subscription:
    put:
      operationId: updateUserSubscription
      tags:
        - Users
      summary: Actualizar suscripción de un usuario (admin)
      description: |
        Cambia el nivel de suscripción de un usuario. Solo para administradores.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            minimum: 1
          description: ID del perfil de usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubscriptionRequest'
      responses:
        '200':
          description: Suscripción actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
security:
  - bearerAuth: []